+++ packages/auth/src/providers/cognito/index.ts
@@ -6,4 +6,5 @@ export { resetPassword } from './apis/resetPassword';
 export { confirmResetPassword } from './apis/confirmResetPassword';
 export { signIn } from './apis/signIn';
 export { resendSignUpCode } from './apis/resendSignUpCode';
+export { confirmSignUp } from './apis/confirmSignUp';
 export { confirmSignIn } from './apis/confirmSignIn';