+++ packages/core/__tests__/clients/middleware/signing/signer/signatureV4/signRequest-test.ts
@@ -4,8 +4,26 @@
 import { signRequest } from '../../../../../../src/clients/middleware/signing/signer/signatureV4/signRequest';
 import { HttpRequest } from '../../../../../../src/clients/types';
 import { signingTestTable } from './testUtils/signingTestTable';
-import { getDefaultRequest, signingOptions } from './testUtils/data';
+import {
+	formattedDates,
+	getDefaultRequest,
+	signingOptions,
+} from './testUtils/data';
 import { SignRequestOptions } from '../../../../../../src/clients/middleware/signing/signer/signatureV4/types';
+import { getSignature } from '../../../../../../src/clients/middleware/signing/signer/signatureV4/utils/getSignature';
+
+jest.mock(
+	'../../../../../../src/clients/middleware/signing/signer/signatureV4/utils/getSignature',
+	() => ({
+		getSignature: jest
+			.fn()
+			.mockImplementation(
+				jest.requireActual(
+					'../../../../../../src/clients/middleware/signing/signer/signatureV4/utils/getSignature'
+				).getSignature
+			),
+	})
+);
 
 describe('signRequest', () => {
 	test.each(
@@ -27,5 +45,25 @@ describe('signRequest', () => {
 		)
 	)('signs request with %s', (_, request, options, expected) => {
 		expect(signRequest(request, options).headers.authorization).toBe(expected);
+	});
+
+	test('should call getSignature with correct parameters', () => {
+		const request = getDefaultRequest();
+		const options = {
+			...signingOptions,
+			uriEscapePath: false,
+		};
+		signRequest(request, options);
+		expect(getSignature).toHaveBeenCalledWith(expect.anything(), {
+			uriEscapePath: false,
+			accessKeyId: options.credentials.accessKeyId,
+			credentialScope: `${formattedDates.shortDate}/${options.signingRegion}/${options.signingService}/aws4_request`,
+			longDate: formattedDates.longDate,
+			secretAccessKey: options.credentials.secretAccessKey,
+			shortDate: formattedDates.shortDate,
+			sessionToken: undefined,
+			signingRegion: options.signingRegion,
+			signingService: options.signingService,
+		});
 	});
 });