+++ packages/storage/__tests__/Storage-unit-test.ts
@@ -1,3 +1,7 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { Credentials } from '@aws-amplify/core';
 import { AWSS3Provider as AWSStorageProvider } from '../src/providers/AWSS3Provider';
 import { Storage as StorageClass } from '../src/Storage';
 import {
@@ -5,9 +9,18 @@ import {
 	Storage as StorageCategory,
 	StorageProvider,
 } from '../src';
-import axios, { CancelToken } from 'axios';
+import { isCancelError } from '../src/AwsClients/S3/utils';
+import { getPrefix } from '../src/common/S3ClientUtils';
+import { AWSS3UploadTask } from '../src/providers/AWSS3UploadTask';
+
+jest.mock('../src/AwsClients/S3/utils');
+jest.mock('../src/common/S3ClientUtils');
+jest.mock('../src/providers/AWSS3UploadTask');
+
+const mockGetPrefix = getPrefix as jest.Mock;
 
 type CustomProviderConfig = {
+	provider: string;
 	foo: boolean;
 	bar: number;
 };
@@ -27,6 +40,11 @@ const options = {
 	level: 'level',
 };
 
+/**
+ * CAUTION: This mock class implements a publically available interface `StorageProvider` which customers can use to
+ * implement custom providers. Exercise caution when modifying this class as additive changes to this interface can
+ * break customers when not marked as optional.
+ */
 class TestCustomProvider implements StorageProvider {
 	getProviderName(): string {
 		return 'customProvider' as const;
@@ -57,6 +75,11 @@ class TestCustomProvider implements StorageProvider {
 	}
 }
 
+/**
+ * CAUTION: This mock class implements a publically available interface `StorageProvider` which customers can use to
+ * implement custom providers. Exercise caution when modifying this class as additive changes to this interface can
+ * break customers when not marked as optional.
+ */
 class TestCustomProviderWithCopy
 	extends TestCustomProvider
 	implements StorageProvider
@@ -64,9 +87,45 @@ class TestCustomProviderWithCopy
 	copy(
 		src: { key: string },
 		dest: { key: string },
-		config: CustomProviderConfig
+		config?: CustomProviderConfig
 	) {
 		return Promise.resolve({ newKey: 'copy' });
+	}
+}
+
+/**
+ * CAUTION: This mock class implements a publically available interface `StorageProvider` which customers can use to
+ * implement custom providers. Exercise caution when modifying this class as additive changes to this interface can
+ * break customers when not marked as optional.
+ */
+class TestCustomProviderWithGetProperties
+	extends TestCustomProvider
+	implements StorageProvider
+{
+	getProperties(key: string, options?: CustomProviderConfig) {
+		return Promise.resolve({ newKey: 'getProperties' });
+	}
+}
+
+/**
+ * CAUTION: This mock class implements a publically available interface `StorageProvider` which customers can use to
+ * implement custom providers. Exercise caution when modifying this class as additive changes to this interface can
+ * break customers when not marked as optional.
+ */
+class TestCustomProviderWithOptionalAPI
+	extends TestCustomProvider
+	implements StorageProvider
+{
+	copy(
+		src: { key: string },
+		dest: { key: string },
+		config?: CustomProviderConfig
+	) {
+		return Promise.resolve({ newKey: 'copy' });
+	}
+
+	getProperties(key: string, options?: CustomProviderConfig) {
+		return Promise.resolve({ newKey: 'getProperties' });
 	}
 }
 
@@ -278,22 +337,64 @@ describe('Storage', () => {
 			});
 		});
 
-		test('vault level is always private', () => {
+		describe('storage value', () => {
 			const storage = StorageCategory;
-			expect.assertions(3);
-			storage.vault.configure = jest.fn().mockImplementation(configure => {
-				expect(configure).toEqual({
-					AWSS3: { bucket: 'bucket', level: 'private', region: 'region' },
-				});
-			});
-			const aws_options = {
-				aws_user_files_s3_bucket: 'bucket',
-				aws_user_files_s3_bucket_region: 'region',
-			};
+			const originalConfigure = storage.vault.configure.bind(storage.vault);
 
-			storage.configure(aws_options);
-			storage.configure({ Storage: { level: 'protected' } });
-			storage.configure({ Storage: { level: 'public' } });
+			afterEach(() => {
+				storage.configure = originalConfigure;
+			});
+
+			it('should always use private access level when configure', () => {
+				expect.assertions(3);
+				storage.vault.configure = jest.fn().mockImplementation(configure => {
+					originalConfigure(configure);
+					expect(configure).toEqual({
+						AWSS3: { bucket: 'bucket', level: 'private', region: 'region' },
+					});
+				});
+				const aws_options = {
+					aws_user_files_s3_bucket: 'bucket',
+					aws_user_files_s3_bucket_region: 'region',
+				};
+
+				storage.configure(aws_options);
+				storage.configure({ Storage: { level: 'protected' } });
+				storage.configure({ Storage: { level: 'public' } });
+			});
+
+			it('should use private access level to initiate multipart upload', () => {
+				const aws_options = {
+					aws_user_files_s3_bucket: 'bucket',
+					aws_user_files_s3_bucket_region: 'region',
+				};
+				storage.configure(aws_options);
+
+				mockGetPrefix.mockImplementationOnce(config => {
+					expect(config).toHaveProperty('level', 'private');
+					return 'fake-prefix';
+				});
+
+				jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+					return new Promise((res, rej) => {
+						res({});
+					});
+				});
+
+				storage.vault.put(
+					'test-key',
+					new Blob([new Uint8Array(1024 * 1024 * 6)]),
+					{
+						resumable: true,
+					}
+				);
+
+				expect(AWSS3UploadTask).toHaveBeenCalledWith(
+					expect.objectContaining({
+						level: 'private',
+					})
+				);
+			});
 		});
 
 		test('normal storage level is public by default', () => {
@@ -587,6 +688,10 @@ describe('Storage', () => {
 			}
 		});
 
+		/**
+		 * Custom providers are a publically available feature via the `StorageProvider` interface. Exercise caution when
+		 * updating these to ensure backwards compatibility.
+		 */
 		test('get with custom provider', async () => {
 			const customProvider = new TestCustomProvider();
 			const customProviderGetSpy = jest.spyOn(customProvider, 'get');
@@ -599,7 +704,11 @@ describe('Storage', () => {
 			expect(customProviderGetSpy).toBeCalled();
 			expect(getRes.newKey).toEqual('get');
 		});
-		// backwards compatible with current custom provider user
+
+		/**
+		 * Custom providers are a publically available feature via the `StorageProvider` interface. Exercise caution when
+		 * updating these to ensure backwards compatibility.
+		 */
 		test('get with custom provider should work with no generic type provided', async () => {
 			const customProvider = new TestCustomProvider();
 			const customProviderGetSpy = jest.spyOn(customProvider, 'get');
@@ -611,6 +720,109 @@ describe('Storage', () => {
 				config3: 'config',
 			});
 			expect(customProviderGetSpy).toBeCalled();
+		});
+	});
+
+	describe('getProperties test', () => {
+		let storage: StorageClass;
+		let provider: StorageProvider;
+		let getPropertiesSpy: jest.SpyInstance;
+
+		beforeEach(() => {
+			storage = new StorageClass();
+			provider = new AWSStorageProvider();
+			storage.addPluggable(provider);
+			storage.configure(options);
+			getPropertiesSpy = jest
+				.spyOn(AWSStorageProvider.prototype, 'getProperties')
+				.mockImplementation(() =>
+					Promise.resolve({
+						contentType: 'text/plain',
+						contentLength: 100,
+						eTag: 'etag',
+						lastModified: new Date('20 Oct 2023'),
+						metadata: { key: '' },
+					})
+				);
+		});
+
+		afterEach(() => {
+			jest.clearAllMocks();
+		});
+
+		describe('with default S3 provider', () => {
+			test('get properties of object successfully', async () => {
+				const result = await storage.getProperties('key');
+				expect(getPropertiesSpy).toBeCalled();
+				expect(result).toEqual({
+					contentType: 'text/plain',
+					contentLength: 100,
+					eTag: 'etag',
+					lastModified: new Date('20 Oct 2023'),
+					metadata: { key: '' },
+				});
+				getPropertiesSpy.mockClear();
+			});
+
+			test('get properties of object with available config', async () => {
+				await storage.getProperties('key', {
+					SSECustomerAlgorithm: 'aes256',
+					SSECustomerKey: 'key',
+					SSECustomerKeyMD5: 'md5',
+				});
+			});
+		});
+
+		test('get properties without provider', async () => {
+			const storage = new StorageClass();
+			try {
+				await storage.getProperties('key');
+			} catch (err) {
+				expect(err).toEqual(new Error('No plugin found with providerName'));
+			}
+		});
+
+		test('get properties with custom provider should work with no generic type provided', async () => {
+			const customProvider = new TestCustomProviderWithGetProperties();
+			const customProviderGetPropertiesSpy = jest.spyOn(
+				customProvider,
+				'getProperties'
+			);
+			storage.addPluggable(customProvider);
+			await storage.getProperties('key', {
+				provider: 'customProvider',
+			});
+			expect(customProviderGetPropertiesSpy).toBeCalled();
+		});
+
+		test('getProperties object with custom provider', async () => {
+			const customProvider = new TestCustomProviderWithGetProperties();
+			const customProviderGetPropertiesSpy = jest.spyOn(
+				customProvider,
+				'getProperties'
+			);
+			storage.addPluggable(customProvider);
+			await storage.getProperties<TestCustomProviderWithGetProperties>('key', {
+				foo: true,
+				bar: 1,
+				provider: 'customProvider',
+			});
+			expect(customProviderGetPropertiesSpy).toBeCalled();
+		});
+
+		test('getProperties object with optionalAPI custom provider', async () => {
+			const customProvider = new TestCustomProviderWithOptionalAPI();
+			const customProviderGetPropertiesSpy = jest.spyOn(
+				customProvider,
+				'getProperties'
+			);
+			storage.addPluggable(customProvider);
+			await storage.getProperties<TestCustomProviderWithOptionalAPI>('key', {
+				foo: true,
+				bar: 1,
+				provider: 'customProvider',
+			});
+			expect(customProviderGetPropertiesSpy).toBeCalled();
 		});
 	});
 
@@ -935,24 +1147,44 @@ describe('Storage', () => {
 			const customProviderWithCopy = new TestCustomProviderWithCopy();
 			const customProviderCopySpy = jest.spyOn(customProviderWithCopy, 'copy');
 			storage.addPluggable(customProviderWithCopy);
-			const copyRes = await storage.copy<TestCustomProviderWithCopy>(
-				{ key: 'src' },
-				{ key: 'dest' },
-				{
-					provider: 'customProvider',
-					foo: false,
-					bar: 40,
-				}
-			);
+			const copyRes: { newKey: string } =
+				await storage.copy<TestCustomProviderWithCopy>(
+					{ key: 'src' },
+					{ key: 'dest' },
+					{
+						provider: 'customProvider',
+						foo: false,
+						bar: 40,
+					}
+				);
 			expect(customProviderCopySpy).toBeCalled();
 			expect(copyRes.newKey).toEqual('copy');
 		});
-		// backwards compatible with current custom provider user
+
+		test('copy object with optionalAPI custom provider', async () => {
+			const customProviderWithCopy = new TestCustomProviderWithOptionalAPI();
+			const customProviderCopySpy = jest.spyOn(customProviderWithCopy, 'copy');
+			storage.addPluggable(customProviderWithCopy);
+			const copyRes: { newKey: string } =
+				await storage.copy<TestCustomProviderWithOptionalAPI>(
+					{ key: 'src' },
+					{ key: 'dest' },
+					{
+						provider: 'customProvider',
+						foo: false,
+						bar: 40,
+					}
+				);
+			expect(customProviderCopySpy).toBeCalled();
+			expect(copyRes.newKey).toEqual('copy');
+		});
+
+		//backwards compatible with current custom provider user
 		test('copy object with custom provider should work with no generic type provided', async () => {
 			const customProviderWithCopy = new TestCustomProviderWithCopy();
 			const customProviderCopySpy = jest.spyOn(customProviderWithCopy, 'copy');
 			storage.addPluggable(customProviderWithCopy);
-			await storage.copy(
+			const copyRes: { newKey: string } = await storage.copy(
 				{ key: 'src' },
 				{ key: 'dest' },
 				{
@@ -963,34 +1195,43 @@ describe('Storage', () => {
 				}
 			);
 			expect(customProviderCopySpy).toBeCalled();
+			expect(copyRes.newKey).toEqual('copy');
 		});
 	});
 
 	describe('cancel test', () => {
-		let isCancelSpy: jest.SpyInstance;
-		let cancelTokenSpy: jest.SpyInstance;
-		let cancelMock: jest.Mock;
-		let tokenMock: jest.Mock;
+		let mockIsCancelError = isCancelError as jest.Mock;
+		let originalAbortController = window.AbortController;
+		let signalAborted = false;
+		let abortError;
+		const mockAbort = jest.fn();
 
 		beforeEach(() => {
-			cancelMock = jest.fn();
-			tokenMock = jest.fn();
-			isCancelSpy = jest.spyOn(axios, 'isCancel').mockReturnValue(true);
-			cancelTokenSpy = jest
-				.spyOn(axios.CancelToken, 'source')
-				.mockImplementation(() => {
-					return {
-						token: tokenMock as unknown as CancelToken,
-						cancel: cancelMock,
-					};
-				});
+			window.AbortController = jest.fn().mockImplementation(function () {
+				return {
+					signal: {
+						aborted: signalAborted,
+					},
+					abort: mockAbort,
+				};
+			});
+			mockAbort.mockImplementation(message => {
+				signalAborted = true;
+				abortError = new Error(message);
+				throw abortError;
+			});
+			mockIsCancelError.mockImplementation(err => err && err === abortError);
 		});
 
 		afterEach(() => {
 			jest.clearAllMocks();
+			window.AbortController = originalAbortController;
+			signalAborted = false;
+			abortError = undefined;
 		});
 
 		test('happy case - cancel upload', async () => {
+			expect.assertions(3);
 			jest.spyOn(AWSStorageProvider.prototype, 'put').mockImplementation(() => {
 				return Promise.resolve({ key: 'new_object' });
 			});
@@ -998,19 +1239,18 @@ describe('Storage', () => {
 			const provider = new AWSStorageProvider();
 			storage.addPluggable(provider);
 			storage.configure(options);
-			const request = storage.put('test.txt', 'test upload');
-			storage.cancel(request, 'request cancelled');
-			expect(cancelTokenSpy).toBeCalledTimes(1);
-			expect(cancelMock).toHaveBeenCalledTimes(1);
 			try {
-				await request;
+				const request = storage.put('test.txt', 'test upload');
+				storage.cancel(request, 'request cancelled');
 			} catch (err) {
-				expect(err).toEqual('request cancelled');
+				expect(mockAbort).toBeCalledTimes(1);
+				expect(err).toMatchObject(new Error('request cancelled'));
 				expect(storage.isCancelError(err)).toBeTruthy();
 			}
 		});
 
 		test('happy case - cancel download', async () => {
+			expect.assertions(3);
 			jest.spyOn(AWSStorageProvider.prototype, 'get').mockImplementation(() => {
 				return Promise.resolve('some_file_content');
 			});
@@ -1018,33 +1258,32 @@ describe('Storage', () => {
 			const provider = new AWSStorageProvider();
 			storage.addPluggable(provider);
 			storage.configure(options);
-			const request = storage.get('test.txt', {
-				download: true,
-			});
-			storage.cancel(request, 'request cancelled');
-			expect(cancelTokenSpy).toHaveBeenCalledTimes(1);
-			expect(cancelMock).toHaveBeenCalledWith('request cancelled');
 			try {
+				const request = storage.get('test.txt', {
+					download: true,
+				});
+				storage.cancel(request, 'request cancelled');
 				await request;
 			} catch (err) {
-				expect(err).toEqual('request cancelled');
+				expect(mockAbort).toBeCalledTimes(1);
+				expect(err).toMatchObject(new Error('request cancelled'));
 				expect(storage.isCancelError(err)).toBeTruthy();
 			}
 		});
 
 		test('happy case - cancel copy', async () => {
+			expect.assertions(3);
 			const storage = new StorageClass();
 			const provider = new AWSStorageProvider();
 			storage.addPluggable(provider);
 			storage.configure(options);
-			const request = storage.copy({ key: 'src' }, { key: 'dest' }, {});
-			storage.cancel(request, 'request cancelled');
-			expect(cancelTokenSpy).toHaveBeenCalledTimes(1);
-			expect(cancelMock).toHaveBeenCalledWith('request cancelled');
 			try {
+				const request = storage.copy({ key: 'src' }, { key: 'dest' }, {});
+				storage.cancel(request, 'request cancelled');
 				await request;
 			} catch (err) {
-				expect(err).toEqual('request cancelled');
+				expect(mockAbort).toBeCalledTimes(1);
+				expect(err).toMatchObject(new Error('request cancelled'));
 				expect(storage.isCancelError(err)).toBeTruthy();
 			}
 		});
@@ -1052,7 +1291,7 @@ describe('Storage', () => {
 		test('isCancelError called', () => {
 			const storage = new StorageClass();
 			storage.isCancelError({});
-			expect(isCancelSpy).toHaveBeenCalledTimes(1);
+			expect(mockIsCancelError).toHaveBeenCalledTimes(1);
 		});
 	});
 });